Avancement de la page discover

+++ plugins/FCC/Functionality/Meta_Boxes_Product.php
@@ -216,11 +216,13 @@ function enqueue()
 	wp_enqueue_style(  'truongwp-gallery-meta-box',  plugins_url() . '/FCC/css/fcc-plugin.css', array(), false );
 	wp_enqueue_script( 'truongwp-gallery-meta-box',  plugins_url() . '/FCC/js/fcc-plugin.js', array( 'backbone', 'jquery' ), false, true );
 
-	wp_enqueue_style( 'bootstrap-4', get_template_directory_uri() . '/css/bootstrap.min.css', false, '4', 'all' );
-	wp_enqueue_style( 'bootstrap-grid', get_template_directory_uri() . '/css/bootstrap-grid.min.css', false, '4', 'all' );
-	wp_enqueue_style( 'bootstrap-reboot', get_template_directory_uri() . '/css/bootstrap-reboot.min.css', false, '4', 'all' );
-	wp_enqueue_script( 'bootstrap-js', get_template_directory_uri() . '/js/bootstrap.min.js', array('jquery'), '1.4.1', true );
-	wp_enqueue_script( 'bootstrap.bundle-js', get_template_directory_uri() . '/js/bootstrap.bundle.min.js', array('jquery'), '1.4.1', true );
+	if (get_current_screen()->post_type == 'product') {
+		wp_enqueue_style( 'bootstrap-4', get_template_directory_uri() . '/css/bootstrap.min.css', false, '4', 'all' );
+		wp_enqueue_style( 'bootstrap-grid', get_template_directory_uri() . '/css/bootstrap-grid.min.css', false, '4', 'all' );
+		wp_enqueue_style( 'bootstrap-reboot', get_template_directory_uri() . '/css/bootstrap-reboot.min.css', false, '4', 'all' );
+		wp_enqueue_script( 'bootstrap-js', get_template_directory_uri() . '/js/bootstrap.min.js', array('jquery'), '1.4.1', true );
+		wp_enqueue_script( 'bootstrap.bundle-js', get_template_directory_uri() . '/js/bootstrap.bundle.min.js', array('jquery'), '1.4.1', true );
+	}
 }
 add_action( 'admin_enqueue_scripts',  'enqueue'  );
